Added checkout form.

+++ src/app/modules/products/components/cart-details/cart-details.component.ts
@@ -35,6 +35,8 @@ export class CartDetailsComponent implements OnInit {
   }
 
   removeItem(item: CartItem){
-    this.cartService.removeItem(item);
+    if (confirm('Chcesz usunąć produkt: ' + item.name + '?')){
+      this.cartService.removeItem(item);
+    }
   }
 }
